requirements.txt
aio-pika==9.4.3
aioch==0.0.2
aiochclient==2.6.0
aiofiles==23.2.1
aiogram==3.13.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiormq==6.8.1
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
asynch==0.2.4
attrs==24.2.0
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
ciso8601==2.3.1
click==8.1.7
clickhouse-connect==0.7.18
clickhouse-driver==0.2.9
contourpy==1.3.0
cycler==0.12.1
dnspython==2.6.1
Faker==28.4.1
fastapi==0.112.1
ffmpy==0.4.0
filelock==3.16.0
Flask==3.0.3
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.9.0
gradio==4.44.0
gradio_client==1.3.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.24.7
Hypercorn==0.17.3
hyperframe==6.0.1
idna==3.7
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.4.7
leb128==1.0.8
lz4==4.3.3
magic-filter==1.0.12
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
multidict==6.1.0
nest-asyncio==1.6.0
numpy==2.1.1
orjson==3.10.7
packaging==24.1
pamqp==3.3.0
pandas==2.2.2
pika==1.3.2
pika-stubs==0.1.3
pillow==10.4.0
priority==2.0.0
pydantic==2.8.2
pydantic_core==2.20.1
pydub==0.25.1
Pygments==2.18.0
pymongo==4.8.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-multipart==0.0.9
python-telegram-bot==21.5
pytz==2024.1
PyYAML==6.0.2
requests==2.32.3
rich==13.8.1
ruff==0.6.4
semantic-version==2.10.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
starlette==0.38.2
tomlkit==0.12.0
tqdm==4.66.5
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.2
uvicorn==0.30.6
websockets==12.0
Werkzeug==3.0.4
wsproto==1.2.0
xmltodict==0.13.0
yarl==1.11.1
zstandard==0.23.0
zstd==1.5.5.1

selenium~=4.24.0
Telethon~=1.36.0